Popadditives fix
Fixed the issues with the ? icon next to fluid additives not working in a very unsatisfactory way. There seems to be some sort of namespace conflict with the "popadditives" div that causes a soft hang - the app carries on but the popover never appears. Adding a layer with a different name solves the problem but doesn't explain why it exists in the first place.